+++ tutorial/02_star.sh
@@ -13,10 +13,9 @@ NAME=(
     "HEK293_siCTCF_rep2"
 )
 
-db=Ensembl
-build=GRCh38
-#sing="singularity exec --bind /work,/work2 /work/SingularityImages/rumball.0.3.0.sif"
+#sing="singularity exec --bind /work,/work2 /work/SingularityImages/rumball.0.4.2.sif"
 sing="singularity exec rumball.sif"
+
 Ddir=Ensembl-GRCh38/
 
 mkdir -p log